feat(deepmerge): add WASM binding via core split
Extracts the JSON-safe deep-merge algorithm into a new internal crate
`_deepmerge-core` (publish = false). The napi crate becomes a thin
wrapper; a new `wasm/` sub-crate exposes `mergeJson` / `mergeAllJson`
via wasm-bindgen + serde-wasm-bindgen.

Also hardens CI workflow:
- ci.yml: explicit top-level `permissions: contents: read`
  (addresses CodeQL finding "Workflow does not contain permissions")
- wasm-test + bundle-size jobs: switch to `taiki-e/install-action` for
  wasm-pack (more reliable than curl init.sh; auto-adds to PATH); add
  pre-flight diagnostics; add `actions/setup-node` to wasm-test for the
  `--node` test runner; strip leading whitespace from `wc -c` output
  before integer comparison
