+++ crates/_deepmerge-core/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "amigo-deepmerge-core"
+version = "0.1.1"
+edition = "2024"
+publish = false
+
+[lib]
+
+[dependencies]
+serde_json = "1"
