+++ crates/deepmerge/wasm/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "amigo-deepmerge-wasm"
+version = "0.1.1"
+edition = "2024"
+publish = false
+
+[lib]
+# rlib needed for wasm-bindgen-test integration tests; cdylib for the
+# actual WASM artifact emitted by wasm-pack.
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+amigo-deepmerge-core = { path = "../../_deepmerge-core" }
+wasm-bindgen = { workspace = true }
+serde-wasm-bindgen = { workspace = true }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+
+[dev-dependencies]
+wasm-bindgen-test = { workspace = true }
+
+[package.metadata.wasm-pack.profile.release]
+# wasm-opt is invoked from a pre-built binaryen binary downloaded by
+# wasm-pack at build time. Hermetic builds disable it; CI runs
+# wasm-opt -Oz as a separate step after wasm-pack build.
+wasm-opt = false
